requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
frozenlist==1.8.0
greenlet==3.3.2
idna==3.11
multidict==6.7.1
numpy==2.4.2
pandas==3.0.1
propcache==0.4.1
python-dateutil==2.9.0.post0
six==1.17.0
SQLAlchemy==2.0.47
typing_extensions==4.15.0
tzdata==2025.3
yarl==1.22.0
python-dotenv
google-generativeai
requests
scikit-learn
optuna
